Move rhs components to a folder
